+++ .vscode/settings.json
@@ -32,6 +32,7 @@
     "orbicularis",
     "pachypodium",
     "riverpod",
+    "ryotaiwamoto",
     "spinkit",
     "unfocus",
     "YYYYMMDDHHMM"
