+++ lib/pages/sign_up/src/sign_up.dart
@@ -1,0 +1,4 @@
+export '../src/widgets/email_text_form.dart';
+export '../src/widgets/password_text_form.dart';
+export '../src/widgets/terms_and_privacy_policy_text.dart';
+export '../src/widgets/user_name_text_form.dart';
