+++ lib/pages/sign_up/src/widgets/email_text_form.dart
