+++ lib/pages/sign_up/src/widgets/password_text_form.dart
